% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactFactors.R
\docType{data}
\name{impactFactors}
\alias{impactFactors}
\title{All the impact factors available for the Waste Impact Calculator
framework}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4096 rows and 11 columns.
}
\source{
\url{https://github.com/OR-Dept-Environmental-Quality/wic-base/blob/main/documentation/technical-overview-of-the-wic.pdf}
For a menu of WIC repos:

\url{https://or-dept-environmental-quality.github.io/wic/}
}
\usage{
impactFactors
}
\description{
For an introduction to WIC, including data dictionaries:
}
\keyword{datasets}
