% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wicImpactFactors.R
\docType{data}
\name{wicImpactFactors}
\alias{wicImpactFactors}
\title{All the impact factors available for the Waste Impact Calculator
framework}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4096 rows and 11 columns.
}
\source{
\url{https://github.com/OR-Dept-Environmental-Quality/wic-base/blob/main/documentation/technical-overview-of-the-wic.pdf}

\url{https://github.com/OR-Dept-Environmental-Quality/wic-base/blob/main/documentation/example-applications-of-the-wic.pdf}

\url{https://or-dept-environmental-quality.github.io/wic/}
}
\usage{
wicImpactFactors
}
\description{
For documentation on WIC, see the items listed under SOURCE below
}
\keyword{datasets}
